Completed SFX Manager
Finished the SFX manager
Attach an "SFX Manager" (found in Main Menu Scene) to any object you want to emit the sounds in our Audio > SFX folder.
As long as you copy/paste the SFX Manager and attatch it to any game object or prefab, you don't have to set up any of the sounds, and the audio will play from that object or prefab in game.